Add weight unit support and decimal weights for loads
Introduced support for weight units (lbs and kg) with localization across the app. Updated loads to allow decimal weight values and adjusted validation, serialization, and UI components accordingly. This improves flexibility and accuracy in weight handling.

+++ spec/channels/loads_channel_spec.rb
@@ -42,8 +42,8 @@ RSpec.describe LoadsChannel do
              expect(payload).to match_json(
                                   slug:        String,
                                   name:        String,
-                                  weight:      0,
-                                  bags_weight: Integer
+                                  weight:      0.0,
+                                  bags_weight: Numeric
                                 )
            end)
   end
@@ -54,8 +54,8 @@ RSpec.describe LoadsChannel do
              expect(payload).to match_json(
                                   slug:        "new-name",
                                   name:        "new name",
-                                  weight:      Integer,
-                                  bags_weight: Integer
+                                  weight:      Numeric,
+                                  bags_weight: Numeric
                                 )
            end)
   end
